Conflicts:
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Controller/CustomerController.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/DAO/Bikes.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/DAO/BikesDAO.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Services/BikesService.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Services/BikesServiceImpl.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Services/CustomerService.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Services/CustomerServiceImpl.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Services/DisposalServiceImpl.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Services/MaintenanceServiceImpl.java
	src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Services/ServicingServiceImpl.java

+++ src/main/java/uk/co/brightfuture/RaysRentals_Balotellitubies/Model/RentalModel.java
@@ -34,7 +34,7 @@ public class RentalModel implements Serializable {
 	private int customer_id;
 	
 	@Column(name = "bike_reference_number")
-	private int bike_reference_number;
+	private Long bike_reference_number;
 	
 	@Column(name = "cost_of_rental")
 	private Double cost_of_rental;
@@ -68,11 +68,11 @@ public class RentalModel implements Serializable {
 		Id = id;
 	}
 
-	public int getBikeReferenceNumber() {
+	public Long getBikeReferenceNumber() {
 		return bike_reference_number;
 	}
 
-	public void setBikeReferenceNumber(int bike_reference_number) {
+	public void setBikeReferenceNumber(Long bike_reference_number) {
 		this.bike_reference_number = bike_reference_number;
 	}
